Test the frame diagnostics, and fix the three defects that found
The family landed untested; deviation 15 said so rather than claiming
otherwise. This closes that gap with 48 tests, 670 -> 718.

Every expectation is derivable by hand from a frame small enough to read,
never captured from a run, and each is chosen so that a plausible wrong
implementation fails it. The percentile boundaries are pinned in both
percentile-bearing types: p0 is the minimum, p100 the maximum, and eight
samples at eight levels queried at p30 separates nearest-rank from the two
definitions that look right next to it. Injecting a floor rank fails two
tests; injecting a strict '>' on the cumulative comparison fails seven.
CountAtOrAbove(255) is the regression test for the byte loop index the
family shipped with, and a loop asserts the cumulative halves partition
the sample at all 255 split points.

Also pinned: clipping as a joint question about a pixel rather than a
marginal one about a channel, at thresholds 255 and 254; that luma and
relative luminance really are different numbers for the same pixel; the
difference bounding box, its first difference in reading order rather than
leftmost, and its empty case; identity answering false where a difference
report throws; stride padding entering neither; and the write/read round
trip through Skia coming back byte-identical.

The allocation claims are now measured through AllocationProbe rather than
asserted in prose: AreIdentical and Between allocate nothing on either
path, a reused FrameStats allocates nothing per frame, and a fresh one
costs the same for 512x512 as for 64x64.

Writing them found three defects, fixed here:

- FrameStats.Percentile computed its rank in double, so 0.07 * 100 being
  7.000000000000001 made the 7th percentile of a hundred pixels return the
  8th. LevelHistogram already used decimal and documented why.
- FrameComparison.Between returned (0, 0, -1, -1) as the empty bounding
  box while its own comment said zeroes and FrameDifference's docs said
  all four were -1. Four zeroes now; the first-difference position stays
  -1, because (0, 0) is a real position.
- DynamicRangeStops documented the 8-bit sRGB ceiling as 12.4 stops. It is
  11.69.

Deviation 15 moves to Implemented and tested, and records the one wrinkle
left: LevelHistogram is public, tested, consumed by nothing, and speaks
percentages where FrameStats speaks quantiles.

+++ src/Najm.Core/Delivery/FrameStats.cs
@@ -243,9 +243,9 @@ public sealed class FrameStats
     /// <remarks>
     /// Absolute black is excluded because almost every frame contains some and it would make the
     /// ratio infinite, saying nothing. What is left is the span the frame actually uses, which for
-    /// an 8-bit sRGB frame is bounded above by about 12.4 stops (level 1 to level 255) — reaching
-    /// that bound is a sign the encoding, not the scene, is the limit. A frame with no lit pixel at
-    /// all reports zero.
+    /// an 8-bit sRGB frame is bounded above by about 11.7 stops — level 1 decodes to 0.000304 of
+    /// full white, and the base-2 logarithm of that ratio is 11.69. Reaching that bound is a sign
+    /// the encoding, not the scene, is the limit. A frame with no lit pixel at all reports zero.
     /// </remarks>
     /// <exception cref="InvalidOperationException">Nothing has been measured yet.</exception>
     public double DynamicRangeStops
@@ -548,6 +548,11 @@ public sealed class FrameStats
     /// between two adjacent 8-bit codes invents a precision the frame does not contain.
     /// </para>
     /// <para>
+    /// The rank is computed in <see cref="decimal"/> rather than <see cref="double"/>, because a
+    /// binary product overshoots an exact integer for ordinary quantiles and the ceiling then asks
+    /// for one rank too many.
+    /// </para>
+    /// <para>
     /// The two ends are the ones implementations get wrong, so they are pinned here:
     /// <c>Percentile(c, 1.0)</c> is exactly <see cref="Maximum(FrameChannel)"/> — a rank of <c>N</c>
     /// selects the last pixel, and an implementation indexing <c>(int)(q · N)</c> off the end or
@@ -571,7 +576,13 @@ public sealed class FrameStats
         }
 
         var pixels = PixelCount;
-        var rank = (long)Math.Ceiling(quantile * pixels);
+
+        // Decimal, not double. The rank is a count, and binary floating point cannot hold most
+        // quantiles exactly: 0.07 * 100 is 7.000000000000001, whose ceiling is 8, so the seventh
+        // percentile of a hundred-pixel frame would quietly answer with the eighth pixel. Decimal
+        // represents the quantiles anyone actually types, so the product is exact for them and the
+        // ceiling means what it says.
+        var rank = (long)Math.Ceiling((decimal)quantile * pixels);
         rank = Math.Clamp(rank, 1L, pixels);
 
         var offset = BucketOffset(channel);
